requirements.txt
anyio==3.7.1
certifi==2023.7.22
click==8.1.6
docutils==0.20.1
exceptiongroup==1.1.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
Jinja2==3.1.2
lockfile==0.12.2
MarkupSafe==2.1.3
nodeenv==1.8.0
paho-mqtt==1.6.1
pip-autoremove==0.10.0
prisma==0.9.1
pydantic==1.10.12
python-daemon==3.0.1
python-dotenv==1.0.0
sniffio==1.3.0
tomlkit==0.12.1
typing_extensions==4.7.1
websockets==11.0.3
